Rename confirmation to message to be more accurate
